+++ src/drivers/Reed_switch_driver.h
@@ -5,10 +5,10 @@
 class Reed_switch_driver
 {
 private:
-    int pin;
+    uint8_t pin;
 
 public:
-    Reed_switch_driver(int _pin)
+    Reed_switch_driver(uint8_t _pin)
     {
         pin = _pin;
         pinMode(_pin, INPUT);
